backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Werkzeug==3.0.1
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.0.1
python-docx==1.1.0
requests==2.31.0
openai==1.12.0